requirements.txt
fastapi
uvicorn
streamlit
google-genai
pandas
joblib
scikit-learn
catboost
requests
skops